requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
cors==1.0.1
Django==4.2.15
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
idna==3.8
pillow==10.4.0
PyJWT==2.9.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-file==2.1.0
setuptools==74.0.0
sqlparse==0.5.1
tldextract==5.1.2
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.7.0
zope.event==5.0
zope.interface==7.0.3
